refactor: ViewHandlers 제거
- Resolver -> ViewResolver 이름 변경
- Resolver가 HttpResponse처리하도록 변경
- 외부 주입
