--- src/main/java/http/view/Resolver.java
@@ -1,7 +1,0 @@
-package http.view;
-
-import http.model.HttpResponse;
-
-public interface Resolver {
-    HttpResponse resolve(ModelAndView modelAndView);
-}
