+++ src/main/java/http/view/TempleteEngineResolver.java
@@ -2,7 +2,7 @@ package http.view;
 
 import http.model.HttpResponse;
 
-public class ModelResolver implements Resolver {
+public class TempleteEngineResolver implements ViewResolver {
     @Override
     public HttpResponse resolve(ModelAndView modelAndView) {
         return null;
